+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG for Sulu
 
 * dev-master
     * HOTFIX      #3912 [ContentBundle]           Fixed content-query when a child is broken
+    * FEATURE     #3914 [ContactBundle]           Added latitude/longitude to address 
     * FEATURE     #3905 [MediaBundle]             Add canonical and robots noIndex headers to download of old versions 
     * ENHANCEMENT #3899 [PreviewBundle]           Replaced links in preview with '#'
     * ENHANCEMENT #3884 [ContentBundle]           Improved developer-experience when overriding content-teaser-provider